+++ plugins/i18n/builds/wtf-plugin-i18n.cjs
@@ -1,4 +1,4 @@
-/* wtf-plugin-i18n 1.0.1  MIT */
+/* wtf-plugin-i18n 1.0.2  MIT */
 (function (global, factory) {
   typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :
   typeof define === 'function' && define.amd ? define(factory) :
@@ -875,6 +875,7 @@
     'cite web': citeWeb,
     'commons cat': commonsCat,
     'start date': startDate,
+    nihongo: ['japonês', 'japonais', 'нихонго']
   };
 
   const plugin = function (_models, templates) {